fix wrong column order for createTableQuery for pg12
Moving from pg9 to pg12, the column order for the createTableQuery
is now in the wrong order. This is probably due to the fact that
array_agg does not guarantee order unless you use an inner ORDER
BY, but it is not clear, but this does fix things at least.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>