replace prepublish with prepublishOnly in package.json
prepublish was deprecated in npm@4 due to it being confusing to users
as it would run both on publishing and installing a package. This task
does not do anything necessary when running npm install for development,
and is only really useful to ensure we have the latest compiled scripts in
the lib/ directory before running npm publish, hence this is changed to
prepublishOnly instead of prepare.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>