Add support for Cassandra 2.1+
The Cassandra driver we use works with both Cassandra 2.x and 3.x,
however, the system schemas are different. To make Sqlectron work
with Cassandra 2.x, we need to send different metadata queries to
older servers.

The approach in this PR is to use the metadata provided by the
driver whenever possible (already implemented in the schema parser
from of the driver). Otherwise, we define two variants of the schema
queries and send one corresponding to the server version.