use iv functions for cipher functions (#84)
This moves sqlectron from using the deprecated and insecure
createCipher and createDecipher methods to createCiperiv and
createDicipheriv. However, sqlectron-core retains backwards
compatibility in that it will still load old-style passwords
and decode them, but will always use the new methods when
encrypting the passwords. The library does not force a downstream
consumer to necessarily upgrade to the newer crypto methods,
but it is heavily encouraged.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>