add version_compare utility function
This provides a simplistic version_compare function, which will
be necessary for follow-up work where need to run different queries
depending on the version the client is running on. This will be done
in conjuction with another PR which adds a uniform "getVersion()"
function to all clients which will provide the same object structure
to make version comparison easy.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>